refractor: move MainPage and rename purpose to PurposePage
